backend/requirements.txt
# Django
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4

# CORS & Environment
django-cors-headers>=4.4
django-environ>=0.11

# Database & Filters
django-filter>=25.0

# Background Tasks (optional, enable if needed)
# celery>=5.4
# redis>=5.0

# Testing
pytest>=8.0
pytest-django>=4.9
